diplom_main/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
colorama==0.4.6
cssselect2==0.8.0
Django==5.2
django-crispy-forms==2.4
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.3.0
fonttools==4.58.4
gunicorn==23.0.0
iniconfig==2.1.0
mccabe==0.7.0
packaging==25.0
phonenumbers==9.0.3
pillow==11.2.1
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pydyf==0.11.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.9.0
pyphen==0.17.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
